==> notebooks/modeling/3.0-models.ipynb <==
import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestClassifier
from sklearn import metrics
import matplotlib.pyplot as plt
import seaborn as sns

data = pd.read_pickle('../../data/processed/all_samples.pickle')

features = ['daylight_yn',
            'holiday_yn',
            'rush_hour_yn',
            'temp',
            'wind_speed',
            'precipitation',
            'road_length',
            'class_freeway',
            'class_local',
            'class_major',
            'class_other',
            'class_unimproved']

labels = 'accident_yn'

X = data[features]
y = data[labels]

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.4, random_state=42)

forest_100 = RandomForestClassifier(n_estimators=100)

forest_100.fit(X_train, y_train)

y_pred_100 = forest_100.predict(X_test)

print('Random Forest (n=100)')

print('Accuracy:', metrics.accuracy_score(y_test, y_pred_100))
print('Precision:', metrics.precision_score(y_test, y_pred_100))
print('Recall:', metrics.recall_score(y_test, y_pred_100))

feature_importance_100 = pd.Series(forest_100.feature_importances_, index=X.columns).sort_values(ascending=False)
print(feature_importance_100)

forest_500 = RandomForestClassifier(n_estimators=500)

forest_500.fit(X_train, y_train)

y_pred_500 = forest_500.predict(X_test)

print('Random Forest (n=500)')

print('Accuracy:', metrics.accuracy_score(y_test, y_pred_500))
print('Precision:', metrics.precision_score(y_test, y_pred_500))
print('Recall:', metrics.recall_score(y_test, y_pred_500))

feature_importance_500 = pd.Series(forest_500.feature_importances_, index=X.columns).sort_values(ascending=False)
print(feature_importance_500)
