==> notebooks/modeling/3.1-additional-features.ipynb <==
import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestClassifier
from sklearn import metrics
import matplotlib.pyplot as plt
import seaborn as sns

data = pd.read_pickle('../../data/processed/all_samples.pickle')
data['datetime'] = pd.to_datetime(data.date)

data['day'] = data.datetime.dt.weekday_name
data = pd.get_dummies(data, prefix='day', columns=['day'])

features = ['hour',
            'daylight_yn',
            'holiday_yn',
            'rush_hour_yn',
            'temp',
            'wind_speed',
            'precipitation',
            'road_length',
            'class_freeway',
            'class_local',
            'class_major',
            'class_other',
            'class_unimproved',
            'day_Monday',
            'day_Tuesday',
            'day_Wednesday',
            'day_Thursday',
            'day_Friday',
            'day_Saturday',
            'day_Sunday']

labels = 'accident_yn'

X = data[features]
y = data[labels]

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.4, random_state=42)

forest = RandomForestClassifier(n_estimators=100)

forest.fit(X_train, y_train)

y_pred = forest.predict(X_test)

print('Random Forest (n=100)')

print('Accuracy:', metrics.accuracy_score(y_test, y_pred))
print('Precision:', metrics.precision_score(y_test, y_pred))
print('Recall:', metrics.recall_score(y_test, y_pred))

feature_importance = pd.Series(forest.feature_importances_, index=X.columns).sort_values(ascending=False)
print(feature_importance)

forest_1k = RandomForestClassifier(n_estimators=1000)

forest_1k.fit(X_train, y_train)

y_pred_1k = forest_1k.predict(X_test)

print('Random Forest (n=0100)')

print('Accuracy:', metrics.accuracy_score(y_test, y_pred_1k))
print('Precision:', metrics.precision_score(y_test, y_pred_1k))
print('Recall:', metrics.recall_score(y_test, y_pred_1k))

feature_importance = pd.Series(forest_1k.feature_importances_, index=X.columns).sort_values(ascending=False)
print(feature_importance)

**n=1,000 results**

*Metrics*

| Metric    | Value              |
| --------- | ------------------ |
| Accuracy  | 0.8289608836950848 |
| Precision | 0.680050627981696  |
| Recall    | 0.5907007425198315 |

*Feature Importance*

| Feature          | Relative Importance |
| ---------------- | ------------------- |
| road_length      | 0.421045            |
| temp             | 0.155326            |
| wind_speed       | 0.099657            |
| class_major      | 0.090212            |
| hour             | 0.081451            |
| class_local      | 0.069730            |
| precipitation    | 0.013593            |
| daylight_yn      | 0.012632            |
| class_freeway    | 0.010737            |
| class_unimproved | 0.008278            |
| holiday_yn       | 0.006628            |
| rush_hour_yn     | 0.005235            |
| day_Monday       | 0.003692            |
| day_Thursday     | 0.003615            |
| day_Sunday       | 0.003610            |
| day_Tuesday      | 0.003590            |
| day_Wednesday    | 0.003531            |
| day_Friday       | 0.003319            |
| day_Saturday     | 0.003096            |
| class_other      | 0.001022            |

important_features = ['hour',
                      'daylight_yn',
                      'temp',
                      'wind_speed',
                      'precipitation',
                      'road_length',
                      'class_freeway',
                      'class_local',
                      'class_major']

X_important = data[important_features]

X_train_imp, X_test_imp, y_train_imp, y_test_imp = train_test_split(X_important, y, test_size=0.4, random_state=42)

forest_imp = RandomForestClassifier(n_estimators=100)

forest_imp.fit(X_train_imp, y_train_imp)

y_pred_imp = forest_imp.predict(X_test_imp)

print('Random Forest (n=100)')

print('Accuracy:', metrics.accuracy_score(y_test_imp, y_pred_imp))
print('Precision:', metrics.precision_score(y_test_imp, y_pred_imp))
print('Recall:', metrics.recall_score(y_test_imp, y_pred_imp))

feature_importance_imp = pd.Series(forest_imp.feature_importances_, index=X_important.columns).sort_values(ascending=False)
print(feature_importance_imp)

most_important_features = ['hour',
                           'temp',
                           'wind_speed',
                           'road_length',
                           'class_local',
                           'class_major']

X_most_important = data[most_important_features]

X_train_most, X_test_most, y_train_most, y_test_most = train_test_split(X_most_important, y, test_size=0.4, random_state=42)

forest_most = RandomForestClassifier(n_estimators=100)

forest_most.fit(X_train_most, y_train_most)

y_pred_most = forest_most.predict(X_test_most)

print('Random Forest (n=100)')

print('Accuracy:', metrics.accuracy_score(y_test_most, y_pred_most))
print('Precision:', metrics.precision_score(y_test_most, y_pred_most))
print('Recall:', metrics.recall_score(y_test_most, y_pred_most))

feature_importance_most = pd.Series(forest_most.feature_importances_, index=X_most_important.columns).sort_values(ascending=False)
print(feature_importance_most)
