==> notebooks/modeling/3.2-random-forest-hyperparameter-tuning.ipynb <==
import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split, RandomizedSearchCV
from sklearn.ensemble import RandomForestClassifier
from sklearn import metrics
import matplotlib.pyplot as plt
import seaborn as sns
from pprint import pprint

data = pd.read_pickle('../../data/processed/all_samples.pickle')
data['datetime'] = pd.to_datetime(data.date)
data['day'] = data.datetime.dt.weekday_name
data = pd.get_dummies(data, prefix='day', columns=['day'])

features = ['hour',
            'daylight_yn',
            'holiday_yn',
            'rush_hour_yn',
            'temp',
            'wind_speed',
            'precipitation',
            'road_length',
            'class_freeway',
            'class_local',
            'class_major',
            'class_other',
            'class_unimproved',
            'day_Monday',
            'day_Tuesday',
            'day_Wednesday',
            'day_Thursday',
            'day_Friday',
            'day_Saturday',
            'day_Sunday']

labels = 'accident_yn'

X = data[features]
y = data[labels]

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.4, random_state=42)

forest = RandomForestClassifier(n_estimators=100, random_state=42)

forest.fit(X_train, y_train)

y_pred = forest.predict(X_test)

print('Random Forest (n=100)')

print('Accuracy:', metrics.accuracy_score(y_test, y_pred))
print('Precision:', metrics.precision_score(y_test, y_pred))
print('Recall:', metrics.recall_score(y_test, y_pred))

feature_importance = pd.Series(forest.feature_importances_, index=X.columns).sort_values(ascending=False)
print(feature_importance)

# Current Parameters
pprint(forest.get_params())

# Random Search

# Number of trees in random forest
n_estimators = [int(x) for x in np.linspace(start=50, stop=150, num=5)]

# Number of features to consider at each split
max_features = ['auto', 'sqrt']

# Maximum number of levels in tree
max_depth = [int(x) for x in np.linspace(start=10, stop=100, num=5)]
max_depth.append(None)

# Minimum number of samples required to split a node
min_samples_split = [2, 5, 10]

# Minimum number of samples required at each leaf node
min_samples_leaf = [1, 2, 4]

# Method of selecting samples for training each tree
bootstrap = [True, False]

random_grid = {'n_estimators': n_estimators,
               'max_features': max_features,
               'max_depth': max_depth,
               'min_samples_split': min_samples_split,
               'min_samples_leaf': min_samples_leaf,
               'bootstrap': bootstrap}

pprint(random_grid)

rf = RandomForestClassifier()

rf_random = RandomizedSearchCV(estimator = rf,
                               param_distributions = random_grid,
                               n_iter=20,
                               n_jobs=-1,
                               cv=3,
                               verbose=2,
                               random_state=42)

rf_random.fit(X_train, y_train)

# save results of 30 min worth of processing...
import pickle
f = open('../../data/processed/rf_random.pickle', 'wb')
pickle.dump(rf_random, f)
f.close()

pprint(rf_random.best_params_)

def evaluate(model, test_features, test_labels):
    predictions = model.predict(test_features)
    accuracy = 100 * metrics.accuracy_score(test_labels, predictions)
    precision = 100 * metrics.precision_score(test_labels, predictions)
    recall = 100 * metrics.recall_score(test_labels, predictions)

    print('Model Performance')
    print('Accuracy:\t{:0.2f}%.'.format(accuracy))
    print('Precision:\t{:0.2f}%.'.format(precision))
    print('Recall:\t{:0.2f}%.'.format(recall))

    return (accuracy, precision, recall)


base_model = RandomForestClassifier(n_estimators=100, random_state=42)
base_model.fit(X_train, y_train)
base_performance = evaluate(base_model, X_test, y_test)

best_random = rf_random.best_estimator_
random_performance = evaluate(best_random, X_test, y_test)

improvement_accuracy = 100 * (random_performance[0] - base_performance[0]) / base_performance[0]
improvement_precision = 100 * (random_performance[1] - base_performance[1]) / base_performance[1]
improvement_recall = 100 * (random_performance[2] - base_performance[2]) / base_performance[2]

print('Accuracy Improvement:\t{:0.2f}%.'.format(improvement_accuracy))
print('Precision Improvement:\t{:0.2f}%.'.format(improvement_precision))
print('Recall Improvement:\t{:0.2f}%.'.format(improvement_recall))

feature_importance_best = pd.Series(best_random.feature_importances_, index=X.columns).sort_values(ascending=False)
print(feature_importance_best)
